backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
pymongo>=4.6.0
motor>=3.3.2

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Validation
email-validator>=2.1.0
pydantic>=2.5.2
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0
certifi>=2024.2.2

# QR Code Generation
qrcode[pil]>=7.4.2
Pillow>=10.1.0

# Email
sendgrid>=6.11.0
jinja2>=3.1.0

# Export
openpyxl>=3.1.2

# Google Auth (optional, if implementing Google Sign-In)
google-auth>=2.23.0
requests>=2.31.0
